generation-service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pillow==11.0.0
httpx==0.28.1
openai==1.55.3
